Harden live loop after full code review
- RPC: transient errors (network, timeouts, 429, 5xx) retry until they
  pass; permanent errors are bounded, so a bad single-block eth_getLogs
  reaches the runner's skip logic instead of stalling forever, and only
  permanent errors split a getLogs range (no request burst on 429)
- Pool cache: bound untrusted fee() to uint24 and ignore it for V2, so a
  fake pool can no longer crash getPool encoding or the SQLite insert
  and make the bot skip whole blocks
- Polling: stay quiet ~10 s after a new block (blocks come every 12 s),
  about 2 eth_blockNumber calls per block instead of about 6
- Multicall: bisect a failing aggregate3 chunk instead of 150 sequential
  single calls
- Telegram sends go through a background queue and no longer block
  block processing
- GC also evicts alerted pools once their cooldown has expired
- Fatal errors (e.g. in replay) are logged through the redacting
  formatter; the supervisor backoff resets after a healthy run

+++ tests/fakechain.py
@@ -57,7 +57,7 @@ class FakeRpc:
         self.fn(address, "token0()", ret(["address"], [token0]))
         self.fn(address, "token1()", ret(["address"], [token1]))
         if fee is not None:
-            self.fn(address, "fee()", ret(["uint24"], [fee]))
+            self.fn(address, "fee()", ret(["uint256"], [fee]))  # как в сыром ответе: 32 байта
 
     def v2_factory(self, factory: str, pairs: dict[tuple[str, str], str]) -> None:
         def get_pair(args: bytes, block: Any) -> bytes:
